Adjusted Player Beam, Added Beam Sound, Changed Beam Sprites, Made Player Attacks Efficient
Player attacks are now much more efficient. Instead of always going through every enemy and checking their hitboxes for attacks, it narrows it down to enemies near and in front of the player.
The beam now has sound, changed range, and new effect. It also drains the players energy slightly when used.
Added a test map for testing purposes.